\name{fitted.spARCH}
\alias{fitted.spARCH}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract model fitted values
}
\description{
The function extracts the fitted values of a spatial ARCH model.
}
\usage{
## S3 method for class 'spARCH'
\method{fitted}{spARCH}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{spARCH} object generated by \code{\link{qml.spARCH}} or \code{\link{qml.SARspARCH}}.}
  \item{...}{
Other arguments.
}
}
\details{
Fitted values extracted from the \code{object}.
}
\author{
Philipp Otto \email{philipp.otto@glasgow.ac.uk}
}
\seealso{
\code{\link[spGARCH]{residuals.spARCH}}, \code{\link[spGARCH]{logLik.spARCH}}.
}
