\name{plot.spARCH}
\alias{plot.spARCH}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Descriptive plots for residuals of a fitted spatial ARCH model
}
\description{
The function depicts several descriptive statistics of the residuals of a fitted spatial ARCH model.
}
\usage{
## S3 method for class 'spARCH'
\method{plot}{spARCH}(x, which = c(1:3), ask, ..., qqline = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{\linkS4class{spARCH} object generated by \code{\link{qml.spARCH}} or \code{\link{qml.SARspARCH}}.}
  \item{which}{
Index number of plot to be returned.
}
  \item{ask}{
if \code{TRUE}, the user is asked before plotting the next figure.
}
  \item{...}{
Other arguments.
}
  \item{qqline}{
A line of the normal distribution for comparison is added to the Q-Q plot, if \code{qqline = TRUE}.
}
}
\details{
The function \code{plot.spARCH} provides several descriptive plots to analyze the residuals of a fitted spatial ARCH model, namely (1) Moran's plot for residuals, (2) Moran's plot for squared residuals, and (3) Normal Q-Q plot for standardized residuals.
}
\author{
Philipp Otto \email{potto@europa-uni.de}
}
\note{
For details of \code{moran.plot} see: \code{help(moran.plot)}.
}
