\name{residuals.spARCH}
\alias{residuals.spARCH}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract model residuals
}
\description{
The function extracts the residuals of a fitted spatial ARCH model.
}
\usage{
## S3 method for class 'spARCH'
\method{residuals}{spARCH}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{\code{spARCH} object generated by \code{\link{qml.spARCH}} or \code{\link{qml.SARspARCH}}.}
  \item{...}{
Other arguments.
}
}
\details{
The function extracts the residuals of a fitted spatial ARCH model.
}
\author{
Philipp Otto \email{philipp.otto@glasgow.ac.uk}
}
\seealso{
\code{\link[spGARCH]{residuals.spARCH}}, \code{\link[spGARCH]{fitted.spARCH}}.
}
