\name{spARCH}
\alias{spARCH}
\alias{spARCH-class}
\docType{class}
\title{Output of Quasi-Maximum-Likelihood Estimation}

\description{
The \code{spARCH} class is a class generated by the estimation functions \code{\link{qml.spARCH}} or \code{\link{qml.SARspARCH}} comprising the results of the quasi-maximum-likelihood estimation.
}

\section{Methods}{
  \describe{
    Several generic methods for summarizing the results are available:
    \item{\code{plot(spARCH_object)}}{\code{\link{plot.spARCH}} - Descriptive plots for residuals of a fitted spatial ARCH model}
    \item{\code{summary(spARCH_object)}}{\code{\link{summary.spARCH}} - The function returns a summary of the model fit of a spatial ARCH model}
    \item{\code{fitted(spARCH_object)}}{\code{\link{fitted.spARCH}} - The function extracts the fitted values of a spatial ARCH model.}
  }
}

